+++ src/qgfOracle/Player.java
@@ -8,7 +8,7 @@ import java.util.UUID;
  *
  */
 public interface Player {
-	record PlayerID(UUID id, long index) {
+	record PlayerID(UUID id, Integer index) {
 		@Override
 		public String toString() {
 			return "P" + index + " (" + id + ")";
